Factor out GraphDescriberManager
